+++ Sources/SwiftUIKit/SHaring/ShareSheet.swift
@@ -1,0 +1,49 @@
+//
+//  ShareScreen.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2020-01-29.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+/**
+ This sheet can be used to present `UIActivityViewController`
+ modals for any kind of shareable content.
+ */
+public struct ShareSheet: UIViewControllerRepresentable {
+    
+    public init(
+        activityItems: [Any],
+        applicationActivities: [UIActivity]? = nil,
+        excludedActivityTypes: [UIActivity.ActivityType]? = nil,
+        callback: Callback? = nil) {
+        self.activityItems = activityItems
+        self.applicationActivities = applicationActivities
+        self.excludedActivityTypes = excludedActivityTypes
+        self.callback = callback
+    }
+    
+    public typealias Callback = (
+        _ activityType: UIActivity.ActivityType?,
+        _ completed: Bool,
+        _ returnedItems: [Any]?,
+        _ error: Error?) -> Void
+      
+    private let activityItems: [Any]
+    private let applicationActivities: [UIActivity]?
+    private let excludedActivityTypes: [UIActivity.ActivityType]?
+    private let callback: Callback?
+      
+    public func makeUIViewController(context: Context) -> UIActivityViewController {
+        let controller = UIActivityViewController(
+            activityItems: activityItems,
+            applicationActivities: applicationActivities)
+        controller.excludedActivityTypes = excludedActivityTypes
+        controller.completionWithItemsHandler = callback
+        return controller
+    }
+      
+    public func updateUIViewController(_ uiViewController: UIActivityViewController, context: Context) {}
+}